Extract UrlCommandTrait
